fix(ci): clear release-mode warnings and materialise disabled-coalescer writes
Release CI builds with RUSTFLAGS="-D warnings"; clear all build/clippy
warnings surfaced in --release and all-targets clippy:

- powerfs-net/transport_rdma: file-level allow for the FFI binding mirror
  (non-camel-case/dead/unused/deprecated/unnecessary-cast mirror the C ABI);
  use io::Error::other instead of ErrorKind::Other
- powerfs-net: drop dead server connection counters, allow excessive-arg
  fns, propagate backoff errors with ?, fix rdma_pingpong example
- powerfs-core/local_fs: bound MutexGuard lifetime, drop unused IO imports
- powerfs-core/volume: remove dead op_counter/opportunistic_flush/flush_option
  (background maintenance thread drives deadline flushing)
- powerfs-filer, powerfs-fuse, powerfs-common: misc clippy/fmt/doc nits

Real bug fix: with the coalescer disabled, record_write returns a forced
flush payload that write_needle_blob discarded with `let _ =`, so the
needle was never materialised and subsequent reads failed with
NeedleNotFound. Materialise it synchronously on both write paths.

Also align two stale tests with current code: async coalescer expiry
semantics and the rdma_buf_num default (32 -> 128, issue #79).

+++ powerfs-filer/src/scrubber.rs
@@ -448,7 +448,7 @@ impl ScrubberWorker {
             }
 
             // 按文件大小选择 EC tier (data_shards, parity_shards)
-            let file_size: u64 = chunks.iter().map(|c| c.size as u64).sum();
+            let file_size: u64 = chunks.iter().map(|c| c.size).sum();
             let (data_shards, parity_shards) = self.config.select_ec_config(file_size);
             let total_shards = (data_shards + parity_shards) as usize;
